+++ server/src/expenses/services/entities.rs
@@ -3,7 +3,7 @@ use serde::Serialize;
 #[derive(Debug, Serialize)]
 pub struct Expense {
     pub id: String,
-    pub amount: f64,
+    pub amount: f32,
     pub description: String,
     pub processed: bool,
     pub date: String,
